+++ src/test/virtual-client.test.ts
@@ -1,7 +1,7 @@
 import { test, describe } from "node:test";
-import { match, strictEqual,  } from "node:assert";
+import { match, strictEqual } from "node:assert";
 import { NiimbotVirtualClient, ResolutionClass } from "..";
-import { B1_V5_22, B21_PRO_V3_09, B21S_V40_28, B2_PRO_V2_01, C1_V3_12, D110_V5_34, D110M_V4_23 } from "./dumps";
+import { B1_V5_22, B21_PRO_V3_09, B21S_V40_28, B2_PRO_V2_01, D110_V5_34, D110M_V4_23 } from "./dumps";
 
 describe("Virtual B1 5.22 test", async () => {
   const client = new NiimbotVirtualClient();
@@ -9,36 +9,47 @@ describe("Virtual B1 5.22 test", async () => {
   await client.connect();
 
   const info = client.getPrinterInfo();
+  const heartbeatInfo = await client.protocol.heartbeat();
 
-  test("modelId", async () => strictEqual(info.modelId!, 4096));
-  test("hardwareVersion", async () => strictEqual(info.hardwareVersion!, "5.10"));
-  test("softwareVersion", async () => strictEqual(info.softwareVersion!, "5.22"));
-  test("printheadWidth", async () => strictEqual(info.printheadWidth!, 384));
-  test("protocolVersion", async () => strictEqual(info.protocolVersion!, 3));
-  test("serial", async () => strictEqual(info.serial!, "G327071185"));
-  test("resolutionClass", async () => strictEqual(info.resolutionClass!, ResolutionClass.DPI203));
+  describe("getPrinterInfo", async () => {
+    test("modelId", async () => strictEqual(info.modelId, 4096));
+    test("hardwareVersion", async () => strictEqual(info.hardwareVersion, "5.10"));
+    test("softwareVersion", async () => strictEqual(info.softwareVersion, "5.22"));
+    test("printheadWidth", async () => strictEqual(info.printheadWidth, 384));
+    test("protocolVersion", async () => strictEqual(info.protocolVersion, 3));
+    test("serial", async () => strictEqual(info.serial, "G327071185"));
+    test("resolutionClass", async () => strictEqual(info.resolutionClass, ResolutionClass.DPI203));
+    test("charge", async () => strictEqual(info.charge, 4));
+  });
+
+  describe("heartbeatInfo", async () => {
+    test("chargeLevel", async () => strictEqual(heartbeatInfo.chargeLevel, 4));
+    test("lidClosed", async () => strictEqual(heartbeatInfo.lidClosed, true));
+  });
 
   await client.disconnect();
 });
 
-describe("Virtual B21 3.09 PRO test", async () => {
+describe("Virtual B21 PRO 3.09 test", async () => {
   const client = new NiimbotVirtualClient();
   client.loadHexDump(B21_PRO_V3_09);
   await client.connect();
 
   const info = client.getPrinterInfo();
 
-  test("modelId", async () => strictEqual(info.modelId!, 785));
-  test("hardwareVersion", async () => strictEqual(info.hardwareVersion!, "3.01"));
-  test("softwareVersion", async () => strictEqual(info.softwareVersion!, "3.09"));
-  test("printheadWidth", async () => strictEqual(info.printheadWidth!, 576));
-  test("protocolVersion", async () => strictEqual(info.protocolVersion!, 5));
-  test("serial", async () => strictEqual(info.serial!, "H613040618"));
-  test("resolutionClass", async () => strictEqual(info.resolutionClass!, ResolutionClass.DPI300));
+  describe("getPrinterInfo", async () => {
+    test("modelId", async () => strictEqual(info.modelId, 785));
+    test("hardwareVersion", async () => strictEqual(info.hardwareVersion, "3.01"));
+    test("softwareVersion", async () => strictEqual(info.softwareVersion, "3.09"));
+    test("printheadWidth", async () => strictEqual(info.printheadWidth, 576));
+    test("protocolVersion", async () => strictEqual(info.protocolVersion, 5));
+    test("serial", async () => strictEqual(info.serial, "H613040618"));
+    test("resolutionClass", async () => strictEqual(info.resolutionClass, ResolutionClass.DPI300));
+    test("charge", async () => strictEqual(info.charge, 2));
+  });
 
   await client.disconnect();
 });
-
 
 describe("Virtual D110 5.34 test", async () => {
   const client = new NiimbotVirtualClient();
@@ -46,17 +57,25 @@ describe("Virtual D110 5.34 test", async () => {
   await client.connect();
 
   const info = client.getPrinterInfo();
+  const heartbeatInfo = await client.protocol.heartbeat();
 
-  test("modelId", async () => strictEqual(info.modelId!, 2304));
-  test("hardwareVersion", async () => match(info.hardwareVersion!, /5\.30/));
-  test("softwareVersion", async () => match(info.softwareVersion!, /5\.34/));
-  test("printheadWidth", async () => strictEqual(info.printheadWidth!, undefined));
-  test("protocolVersion", async () => strictEqual(info.protocolVersion!, 1));
-  test("serial", async () => strictEqual(info.serial!, "G326030306"));
+  describe("getPrinterInfo", async () => {
+    test("modelId", async () => strictEqual(info.modelId, 2304));
+    test("hardwareVersion", async () => match(info.hardwareVersion!, /5\.30/));
+    test("softwareVersion", async () => match(info.softwareVersion!, /5\.34/));
+    test("printheadWidth", async () => strictEqual(info.printheadWidth, undefined));
+    test("protocolVersion", async () => strictEqual(info.protocolVersion, 1));
+    test("serial", async () => strictEqual(info.serial, "G326030306"));
+    test("charge", async () => strictEqual(info.charge, 3));
+  });
+
+  describe("heartbeatInfo", async () => {
+    test("chargeLevel", async () => strictEqual(heartbeatInfo.chargeLevel, 3));
+    test("lidClosed", async () => strictEqual(heartbeatInfo.lidClosed, true));
+  });
 
   await client.disconnect();
 });
-
 
 describe("Virtual D110M 4.23 test", async () => {
   const client = new NiimbotVirtualClient();
@@ -64,18 +83,26 @@ describe("Virtual D110M 4.23 test", async () => {
   await client.connect();
 
   const info = client.getPrinterInfo();
+  const heartbeatInfo = await client.protocol.heartbeat();
 
-  test("modelId", async () => strictEqual(info.modelId!, 2320));
-  test("hardwareVersion", async () => strictEqual(info.hardwareVersion!, "4.01"));
-  test("softwareVersion", async () => strictEqual(info.softwareVersion!, "4.23"));
-  test("protocolVersion", async () => strictEqual(info.protocolVersion!, 4));
-  test("serial", async () => strictEqual(info.serial!, "H322062548"));
-  test("printheadWidth", async () => strictEqual(info.printheadWidth!, 96));
-  test("resolutionClass", async () => strictEqual(info.resolutionClass!, ResolutionClass.DPI203));
+  describe("getPrinterInfo", async () => {
+    test("modelId", async () => strictEqual(info.modelId, 2320));
+    test("hardwareVersion", async () => strictEqual(info.hardwareVersion, "4.01"));
+    test("softwareVersion", async () => strictEqual(info.softwareVersion, "4.23"));
+    test("protocolVersion", async () => strictEqual(info.protocolVersion, 4));
+    test("serial", async () => strictEqual(info.serial, "H322062548"));
+    test("printheadWidth", async () => strictEqual(info.printheadWidth, 96));
+    test("resolutionClass", async () => strictEqual(info.resolutionClass, ResolutionClass.DPI203));
+    test("charge", async () => strictEqual(info.charge, 4));
+  });
+
+  describe("heartbeatInfo", async () => {
+    test("chargeLevel", async () => strictEqual(heartbeatInfo.chargeLevel, 4));
+    test("lidClosed", async () => strictEqual(heartbeatInfo.lidClosed, true));
+  });
 
   await client.disconnect();
 });
-
 
 describe("Virtual B2 PRO 2.09 test", async () => {
   const client = new NiimbotVirtualClient();
@@ -83,14 +110,23 @@ describe("Virtual B2 PRO 2.09 test", async () => {
   await client.connect();
 
   const info = client.getPrinterInfo();
+  const heartbeatInfo = await client.protocol.heartbeat();
 
-  test("modelId", async () => strictEqual(info.modelId!, 6912));
-  test("hardwareVersion", async () => strictEqual(info.hardwareVersion!, "2.01"));
-  test("softwareVersion", async () => strictEqual(info.softwareVersion!, "2.09"));
-  test("protocolVersion", async () => strictEqual(info.protocolVersion!, 5));
-  test("serial", async () => strictEqual(info.serial!, "I122050127"));
-  test("printheadWidth", async () => strictEqual(info.printheadWidth!, 576));
-  test("resolutionClass", async () => strictEqual(info.resolutionClass!, ResolutionClass.DPI300));
+  describe("getPrinterInfo", async () => {
+    test("modelId", async () => strictEqual(info.modelId, 6912));
+    test("hardwareVersion", async () => strictEqual(info.hardwareVersion, "2.01"));
+    test("softwareVersion", async () => strictEqual(info.softwareVersion, "2.09"));
+    test("protocolVersion", async () => strictEqual(info.protocolVersion, 5));
+    test("serial", async () => strictEqual(info.serial, "I122050127"));
+    test("printheadWidth", async () => strictEqual(info.printheadWidth, 576));
+    test("resolutionClass", async () => strictEqual(info.resolutionClass, ResolutionClass.DPI300));
+    test("charge", async () => strictEqual(info.charge, 60)); // todo: why this range?
+  });
+
+  describe("heartbeatInfo", async () => {
+    test("chargeLevel", async () => strictEqual(heartbeatInfo.chargeLevel, 60));
+    test("lidClosed", async () => strictEqual(heartbeatInfo.lidClosed, true));
+  });
 
   await client.disconnect();
 });
@@ -101,14 +137,22 @@ describe("Virtual B21S 40.28 test", async () => {
   await client.connect();
 
   const info = client.getPrinterInfo();
+  const heartbeatInfo = await client.protocol.heartbeat();
 
-  test("modelId", async () => strictEqual(info.modelId!, 777));
-  test("hardwareVersion", async () => match(info.hardwareVersion!, /40\.10/));
-  test("softwareVersion", async () => match(info.softwareVersion!, /40\.28/));
-  test("protocolVersion", async () => strictEqual(info.protocolVersion!, 0));
-  test("serial", async () => strictEqual(info.serial!, "G626070087"));
-  test("printheadWidth", async () => strictEqual(info.printheadWidth!, undefined));
-  test("resolutionClass", async () => strictEqual(info.resolutionClass!, undefined));
+  describe("getPrinterInfo", async () => {
+    test("modelId", async () => strictEqual(info.modelId, 777));
+    test("hardwareVersion", async () => match(info.hardwareVersion!, /40\.10/));
+    test("softwareVersion", async () => match(info.softwareVersion!, /40\.28/));
+    test("protocolVersion", async () => strictEqual(info.protocolVersion, 0));
+    test("serial", async () => strictEqual(info.serial, "G626070087"));
+    test("printheadWidth", async () => strictEqual(info.printheadWidth, undefined));
+    test("resolutionClass", async () => strictEqual(info.resolutionClass, undefined));
+  });
+
+  describe("heartbeatInfo", async () => {
+    test("chargeLevel", async () => strictEqual(heartbeatInfo.chargeLevel, 3));
+    test("lidClosed", async () => strictEqual(heartbeatInfo.lidClosed, true));
+  });
 
   await client.disconnect();
 });